refactored project
