+++ src/main/java/backup/system/Main.java
@@ -1,21 +1,23 @@
 package backup.system;
 
+import backup.system.drive.handler.Drive;
 import backup.system.file.handler.ConfigData;
-import backup.system.file.handler.FileConfiguration;
+
 import java.util.UUID;
 
 public class Main {
 
     public static void main(String[] args) throws Exception {
-//        FileConfiguration fileConfiguration = new FileConfiguration();
-//        ConfigData configData = new ConfigData();
-//        UUID uuid = UUID.randomUUID();
-//        configData.setUuid(uuid.toString());
-//        configData.setLastModified(2123132141);
-//        fileConfiguration.writeFile(configData);
-//
+        ConfigData configData = new ConfigData();
+        Drive drive = new Drive();
+        UUID uuid = UUID.randomUUID();
+        configData.setUuid(uuid.toString());
+        configData.setLastModified(2123132141);
+//        configData.createConfigData();
+
 //        configData = fileConfiguration.loadFile();
 //        System.out.println(configData.getUuid());
 //        System.out.println(configData.getLastModified());
+
     }
 }
