+++ src/main/java/backup/system/backup/prozess/Visitor.java
@@ -1,0 +1,34 @@
+package backup.system.backup.prozess;
+
+import java.io.IOException;
+import java.nio.*;
+import java.nio.file.FileVisitResult;
+import java.nio.file.FileVisitor;
+import java.nio.file.attribute.BasicFileAttributes;
+
+public class Visitor implements FileVisitor {
+
+    @Override
+    public FileVisitResult preVisitDirectory(Object dir, BasicFileAttributes attrs) throws IOException {
+        return null;
+    }
+
+    @Override
+    public FileVisitResult visitFile(Object file, BasicFileAttributes attrs) throws IOException {
+        return null;
+    }
+
+    @Override
+    public FileVisitResult visitFileFailed(Object file, IOException exc) throws IOException {
+        return null;
+    }
+
+    @Override
+    public FileVisitResult postVisitDirectory(Object dir, IOException exc) throws IOException {
+        return null;
+    }
+
+    public Object startVisitor(){
+        return null;
+    }
+}
