--- src/main/java/backup/system/drive/handler/Find_Drives.java
@@ -1,15 +1,0 @@
-package backup.system.drive.handler;
-
-import java.util.UUID;
-
-public class Find_Drives {
-    // - Laufwerke[]
-
-    /**
-     * Füllt das Array mit gefundenen Laufwerken
-     * @param uuid
-     */
-    public void startSearch(UUID uuid){
-
-    }
-}
