+++ src/main/java/backup/system/uuid/handler/Get_UUID.java
@@ -1,15 +1,28 @@
 package backup.system.uuid.handler;
 
+import backup.system.file.handler.ConfigData;
+
 import java.util.UUID;
 
 public class Get_UUID {
-    
+
     /**
-     * Sucht ob es eine UUID gibt
+     *
+     * @param UUID
      * @return
      */
-    public UUID getUuid(){
-        return null;
+    public boolean compareUUID(String UUID){
+        ConfigData configData = new ConfigData();
+        if(configData.getUuid().equals(UUID)){
+            return true;
+        }
+        return false;
     }
 
+    /**
+     * Füllt das Array mit gefundenen Laufwerken
+     */
+    public void assign(){
+
+    }
 }
